Fix a small error

+++ plugins/stream-debugger/src/main/java/com/intellij/debugger/streams/trace/impl/handler/type/GenericTypeUtil.java
@@ -67,7 +67,7 @@ public class GenericTypeUtil {
     }
 
     if (type.equals(GenericType.OPTIONAL_LONG)) {
-      return GenericType.DOUBLE;
+      return GenericType.LONG;
     }
 
     if (type.equals(GenericType.OPTIONAL_DOUBLE)) {
